fix(server,client): reject unknown args, fail on 2xx non-JSON, keep error body, drop create-side description cap
Four hardening changes, each guarding a class of incident that showed up
as 'success: true but nothing landed':

- server.call_tool: unknown argument keys now return INVALID_ARGS with the
  allowed key list, instead of being silently dropped (e.g. tss vs
  tss_planned). Includes a static test that every handler only reads keys
  present in its schema.
- client._handle_response: HTTP 200/201 with a non-empty body that is not
  JSON (login redirect, WAF/Cloudflare challenge) is now a failure with
  error_code UNPARSEABLE_RESPONSE. Empty body stays success.
- client._handle_response: generic API errors carry the first 300 chars of
  the response body so TP's own validation message reaches the caller.
- CreateWorkoutInput: description no longer capped at 2000 chars (update
  never was; TP has no such cap) and duration_minutes accepts float,
  matching UpdateWorkoutInput. Schema type aligned to number.

+++ src/tp_mcp/tools/workouts.py
@@ -384,7 +384,7 @@ async def tp_create_workout(
     date_str: str,
     sport: str,
     title: str,
-    duration_minutes: int | None = None,
+    duration_minutes: float | None = None,
     description: str | None = None,
     distance_km: float | None = None,
     tss_planned: float | None = None,
